requirements.txt

viam-sdk==0.57.0

typing-extensions
